Revert "chore(build): Rename gulpfile.js to ES6-friendly gulpfile.babel.js"
This reverts commit 547d530d9d3f426adf7193b33a34defaa0b32aa0.
