+++ .gitignore
@@ -1,6 +1,5 @@
 tmp/
 .env
-*.log
 *.sw?
 *.pyc
 .DS_Store
